+++ source/module_lr/ao_to_mo_transformer/ao_to_mo_parallel.cpp
@@ -1,5 +1,5 @@
 #ifdef __MPI
-#include "AX.h"
+#include "ao_to_mo.h"
 #include "module_base/scalapack_connector.h"
 #include "module_base/tool_title.h"
 #include "module_lr/utils/lr_util.h"
