+++ source/module_lr/ao_to_mo_transformer/test/CMakeLists.txt
@@ -1,0 +1,6 @@
+remove_definitions(-DUSE_LIBXC)
+AddTest(
+  TARGET ao_to_mo_test
+  LIBS parameter base ${math_libs} container device psi 
+  SOURCES ao_to_mo_test.cpp ../../utils/lr_util.cpp ../ao_to_mo_parallel.cpp ../ao_to_mo_serial.cpp
+)
